Check if Git LFS is installed
