+++ ersilia/__init__.py
@@ -5,6 +5,8 @@ del _version
 
 # External imports
 import os
+import subprocess as s
+import sys
 
 # Disable GPU
 os.environ["CUDA_VISIBLE_DEVICES"] = "-1"
@@ -65,5 +67,17 @@ def check_install_status():
 
 INSTALL_STATUS = check_install_status()["status"]
 
+# Check if Git LFS is installed
+def check_gitlfs_installed():
+    try:
+        check = s.run(["git-lfs"], capture_output=True)
+        
+    except ModuleNotFoundError:
+        sys.exit('\nGit LFS is not installed. We recommend installing Git LFS to'\
+            ' use large size models.\n\nCheck out https://git-lfs.github.com/'\
+            ' on how to install. After installation, simply use the command'\
+            ' `git lfs install` in your repository.\n')
+
+check_gitlfs_installed()
 
 __all__ = ["__version__"]
